Renamed program message type by executable type. This should break existing programs.
